+++ src/components/Badge/Badge.vue
@@ -1,17 +1,14 @@
 <template>
   <div class="inline-block">
     <div
-      :class="['flex justify-center align-center',
-               { 'px-4 py-2 rounded-full': rounded },
-               { 'px-2 py-1 rounded-sm': square },
-               { 'bg-white-300 text-gray-500': info },
-               { 'bg-mint-100 text-mint-900': success },
-               { 'bg-lemon-100 text-lemon-900': warning },
-               { 'bg-coral-100 text-error': error },
-               { 'bg-turquoise-100 text-turquoise-900': turquoise },
-               { 'bg-flint-100 text-gray-700': flint },
-               { 'bg-gradient-114 from-[#1876db] to-[#5748ff] text-white': gradient },
-               { '!py-0.5': small }
+      :class="['flex justify-center align-center px-4 py-1 rounded-full text-gray-500 bg-gray-50',
+               [ small ? '!type-xs-600' : 'type-small-600' ],
+               { '!bg-white text-gray-500 border border-gray-100': secondary },
+               { 'bg-blue-50 text-blue-600': info },
+               { 'bg-green-50 text-green-700': success },
+               { 'bg-orange-50 text-orange-600': warning },
+               { 'bg-red-50 text-red-600': error },
+               { 'bg-gradient-114 from-[#1876db] to-[#5748ff] !text-white': gradient }
       ]"
     >
       <slot />
@@ -25,16 +22,9 @@ export default {
   props: {
     variant: {
       type: String,
-      default: 'info',
+      default: 'default',
       validator: function (value) {
-        return ['info', 'success', 'warning', 'error', 'turquoise', 'flint', 'gradient'].includes(value);
-      }
-    },
-    shape: {
-      type: String,
-      default: 'rounded',
-      validator: function (value) {
-        return ['rounded', 'square'].includes(value);
+        return ['default', 'secondary', 'info', 'success', 'warning', 'error', 'gradient'].includes(value);
       }
     },
     size: {
@@ -46,6 +36,9 @@ export default {
     }
   },
   computed: {
+    secondary () {
+      return this.variant === 'secondary';
+    },
     info () {
       return this.variant === 'info';
     },
@@ -58,20 +51,8 @@ export default {
     error () {
       return this.variant === 'error';
     },
-    turquoise () {
-      return this.variant === 'turquoise';
-    },
-    flint () {
-      return this.variant === 'flint';
-    },
     gradient () {
       return this.variant === 'gradient';
-    },
-    rounded () {
-      return this.shape === 'rounded';
-    },
-    square () {
-      return this.shape === 'square';
     },
     small () {
       return this.size === 'small';